Add the splash page
This commit includes the classes and resources necessary for creating
the first version splash page. For now the images will just be in a flat
directory in the resources directory, but there may be a better home
for them later.

I had a look at using declarative designs from Vaadin, but I feel
as though they are more overhead than they're worth for our relatively
simple GUI. So I've decided to stick with pure Java declarations.