Add a Dashboard page prototype
This commit adds the next page in the GUI prototype.

Things to note: I've started using the double brace initialisation feature
in Java to give some visual structure to the GUI layout code (it shouldn't
be dangerous in this context). I've also figured out (finally) how to include
third-party addons in Vaadin and have added the vaadin-maven-plugin to help
with widgetset compilation and deployment.

The general steps to adding an add-on are:
* Add the add-on's dependency to maven
* Run the goal `mvn vaadin:update-widgetset`. This adds the widgetset as a
depdency to the WidgetSet.gwt.xml file.
* Run the goal `mvn vaadin:compile`, which generates the client-side widgets
* Run spring boot to see the widget in action!