Update ER diagram to reflect new naming conventions and fix test_suite ownership references
Changed 'version' to 'run', 'test_run' to 'execution' and 'execution' and 'execution_step' to
'test_case_execution' and 'test_step_execution' respectively. This naming convention makes
a bit more sense and brings the docs into line with the initial commit of the domain models.

This commit also fixes up some bad logic with the test_suite concept. Initially, they had a
one-to-many relationship with the 'run' entity, but that doesn't make sense, since we would
want them to be remembered on a per-project basis. Secondly, each run my make use of several
test_suites, and each test_suite may be used by many runs, so now a many-to-many relation
exists there. Finally, test_cases may be included individually to a run along with test_suites
so I've added an analogous many-to-many relationship between test_cases and runs.